requirements.txt
google-genai
opencv-python
pyaudio
pillow
mss
streamlit
streamlit-webrtc
python-dotenv
numpy